requirements.txt
flask==3.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
ta==0.11.0
scikit-learn==1.3.2
prophet==1.1.5
aiohttp==3.9.1
psutil==5.9.6
youtube-transcript-api==0.6.1
feedparser==6.0.10
google-generativeai==0.3.2
groq==0.4.1
huggingface-hub==0.19.4
openai==1.6.1
tenacity==8.2.3
cachetools==5.3.2

# Security & Production
flask-httpauth==4.8.0
flask-limiter==3.5.0
flask-cors==4.0.0
gunicorn==21.2.0
